Move ClusterExecutorImpl exclusive methods to ClusterExecutor interface, plus moving the additional interface's dependencies to its corresponding package
