+++ modules/apps/portal/portal-cluster-multiple/src/test/java/com/liferay/portal/cluster/multiple/internal/TestClusterChannelFactory.java
@@ -5,6 +5,9 @@
 
 package com.liferay.portal.cluster.multiple.internal;
 
+import com.liferay.portal.kernel.cluster.ClusterChannel;
+import com.liferay.portal.kernel.cluster.ClusterReceiver;
+
 import java.net.InetAddress;
 import java.net.NetworkInterface;
 import java.net.SocketException;
