+++ portal-kernel/src/com/liferay/portal/kernel/cluster/ClusterReceiver.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.cluster.multiple.internal;
+package com.liferay.portal.kernel.cluster;
 
 import com.liferay.portal.kernel.cluster.Address;
