+++ portal-kernel/src/com/liferay/portal/kernel/cluster/packageinfo
@@ -1,1 +1,1 @@
-version 9.0.0
+version 10.0.0
